requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pyyaml>=6.0

# EEG processing
mne>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0

# Optional: for .mat files (BCI-IV-2a labels)
# h5py>=3.8.0
